+++ Nos/Models/Author+CoreDataClass.swift
@@ -52,6 +52,13 @@ public class Author: NosManagedObject {
         return nil
     }
     
+    class func find(named name: String, context: NSManagedObjectContext) throws -> [Author] {
+        let fetchRequest = NSFetchRequest<Author>(entityName: String(describing: Author.self))
+        fetchRequest.predicate = NSPredicate(format: "name CONTAINS[cd] %@ OR displayName CONTAINS[cd] %@", name, name)
+        let authors = try context.fetch(fetchRequest)
+        return authors
+    }
+    
     @discardableResult
     class func findOrCreate(by pubKey: HexadecimalString, context: NSManagedObjectContext) throws -> Author {
         if let author = try? Author.find(by: pubKey, context: context) {