+++ Nos/Views/ThreadView.swift
@@ -34,7 +34,7 @@ struct ThreadView: View {
     
     var body: some View {
         LazyVStack {
-            NoteButton(note: root)
+            NoteButton(note: root, isInThreadView: true)
                 .padding(.horizontal)
             ForEach(thread) { event in
                 VStack {
@@ -45,7 +45,7 @@ struct ThreadView: View {
                         }
                         .stroke(style: StrokeStyle(lineWidth: 4, lineCap: .round))
                         .fill(Color.secondaryTxt)
-                        NoteButton(note: event)
+                        NoteButton(note: event, isInThreadView: true)
                             .padding(.horizontal)
                     }
                 }